fix(LP64, 2038): cast time_t vals to `long` instead of `int`
* use %ld for the corresponding format string params [%]

On LP64 platforms, both `time_t` & `long` are 64-bit,
so this prevents any data losts.

On ILP32 platforms, however, `long` is 32-bit,
so this does not introduce any overheads.

* maple/bbsd.c:
    * blog_pid() [%]
* maple/xchatd.c:
    * cuser_free() [%]
* so/bbslua.c:
    * bl_clock() (#ifdef _WIN32)
* util/counter.c:
    * main() count->samehour_max_time [%]
* util/reaper.c:
    * reaper() `life` [%]
    * main() `start` [%]

+++ maple/bbsd.c
@@ -92,7 +92,7 @@ blog_pid(
     if (!msg)
     {
         msg = data;
-        sprintf(data, "Stay: %d (%d)", (int)(time(NULL) - ap_start) / 60, pid);
+        sprintf(data, "Stay: %ld (%d)", (long)(time(NULL) - ap_start) / 60, pid);
     }
     logger_tag(&blog_tlogger, mode, msg);
 }
